feat: splice createServerClient into utils
