+++ server/app/root.tsx
@@ -22,12 +22,8 @@ import Navbar from "~/components/navbar";
 import Footer from "~/components/footer";
 import { json } from "@remix-run/node";
 import { useEffect, useState } from "react";
-import {
-  createBrowserClient,
-  createServerClient,
-  parse,
-  serialize,
-} from "@supabase/ssr";
+import { createBrowserClient } from "@supabase/ssr";
+import { createServerClient } from "~/utils/supabase.server";
 
 export const links: LinksFunction = () => [
   { rel: "stylesheet", href: stylesheet },
@@ -45,26 +41,7 @@ export const loader: LoaderFunction = async ({ request }) => {
     SUPABASE_ANON_KEY: process.env.SUPABASE_ANON_KEY,
   };
 
-  const cookies = parse(request.headers.get("Cookie") ?? "");
-  const headers = new Headers();
-
-  const supabase = createServerClient(
-    process.env.SUPABASE_URL ?? "",
-    process.env.SUPABASE_ANON_KEY ?? "",
-    {
-      cookies: {
-        get(key) {
-          return cookies[key];
-        },
-        set(key, value, options) {
-          headers.append("Set-Cookie", serialize(key, value, options));
-        },
-        remove(key, options) {
-          headers.append("Set-Cookie", serialize(key, "", options));
-        },
-      },
-    }
-  );
+  const { supabase, headers } = createServerClient(request);
 
   const {
     data: { session },
