+++ server/app/utils/supabase.server.ts
@@ -1,0 +1,30 @@
+import {
+  createServerClient as createSupabaseServerClient,
+  parse,
+  serialize,
+} from "@supabase/ssr";
+
+export const createServerClient = (request: Request) => {
+  const cookies = parse(request.headers.get("Cookie") ?? "");
+  const headers = new Headers();
+
+  const supabase = createSupabaseServerClient(
+    process.env.SUPABASE_URL ?? "",
+    process.env.SUPABASE_ANON_KEY ?? "",
+    {
+      cookies: {
+        get(key) {
+          return cookies[key];
+        },
+        set(key, value, options) {
+          headers.append("Set-Cookie", serialize(key, value, options));
+        },
+        remove(key, options) {
+          headers.append("Set-Cookie", serialize(key, "", options));
+        },
+      },
+    }
+  );
+
+  return { supabase, headers };
+};
